+++ changelog.md
@@ -1,7 +1,10 @@
-* unreleased
+* v0.9.3.0 2021-01-02
     - Drop support for GHC < 8.4, require base > 4.11.
     - Bump `cabal-version` to 2.4.
     - Add `Ord` instance for `PlaylistName`, `Path`, and `Value`
+    - Add support for MPD 0.21 style filters to the `Query` type,
+      see [Filters](https://www.musicpd.org/doc/html/protocol.html#filters) in the MPD protocol documentation. 
+      Implemented through the added combinators `/=?`, `%?`, `~?`, `/~?`, `qNot`, `qModSince`, `qFile`, `qBase`.
     
 * v0.9.2.0 2020-10-02
     - New command: `seekCur`